added plot.py runner to handle all plotting and visualizations
